+++ test/CMakeLists.txt
@@ -93,6 +93,7 @@ add_subdirectory(util)
 set(MULTI_TEST_SRC
     autodiff.in.cpp
     backend_binary_elementwise.in.cpp
+    backend_broadcast.in.cpp
     backend_comparison.in.cpp
     backend_reduce.in.cpp
     backend_test.in.cpp
