+++ test/backend_test.in.cpp
@@ -1257,83 +1257,6 @@ NGRAPH_TEST(${BACKEND_NAME}, constant_multi_use)
     EXPECT_EQ(read_vector<int>(r2), std::vector<int>{388});
 }
 
-NGRAPH_TEST(${BACKEND_NAME}, constant_broadcast)
-{
-    const string js =
-        R"([{
-       "name" : "Function_0",
-       "ops" : [
-           {
-             "element_type" :
-                 {"bitwidth" : 32, "c_type_string" : "float", "is_real" : true, "is_signed" : true, "is_quantized" : false},
-             "inputs" : [],
-             "name" : "Parameter_4",
-             "op" : "Parameter",
-             "outputs" : ["Parameter_4"],
-             "shape" : [ 3, 4 ]
-           },
-           {
-             "element_type" :
-                 {"bitwidth" : 32, "c_type_string" : "float", "is_real" : true, "is_signed" : true, "is_quantized" : false},
-             "inputs" : [],
-             "name" : "Parameter_0",
-             "op" : "Parameter",
-             "outputs" : ["Parameter_0"],
-             "shape" : [ 3, 4 ]
-           },
-           {
-             "element_type" :
-                 {"bitwidth" : 32, "c_type_string" : "float", "is_real" : true, "is_signed" : true, "is_quantized" : false},
-             "inputs" : [],
-             "name" : "Constant_1",
-             "op" : "Constant",
-             "outputs" : ["Constant_1"],
-             "shape" : [],
-             "value" : ["0"]
-           },
-           {
-             "axes" : [ 0, 1 ],
-             "element_type" :
-                 {"bitwidth" : 32, "c_type_string" : "float", "is_real" : true, "is_signed" : true, "is_quantized" : false},
-             "inputs" : ["Constant_1"],
-             "name" : "Broadcast_2",
-             "op" : "Broadcast",
-             "outputs" : ["Broadcast_2"],
-             "shape" : [ 3, 4 ]
-           },
-           {
-             "element_type" :
-                 {"bitwidth" : 32, "c_type_string" : "float", "is_real" : true, "is_signed" : true, "is_quantized" : false},
-             "inputs" : [ "Parameter_0", "Broadcast_2" ],
-             "name" : "Maximum_3",
-             "op" : "Maximum",
-             "outputs" : ["Maximum_3"]
-           },
-           {
-             "element_type" :
-                 {"bitwidth" : 32, "c_type_string" : "float", "is_real" : true, "is_signed" : true, "is_quantized" : false},
-             "inputs" : [ "Maximum_3", "Parameter_4" ],
-             "name" : "Multiply_5",
-             "op" : "Multiply",
-             "outputs" : ["Multiply_5"]
-           }
-       ],
-       "parameters" : [ "Parameter_0", "Parameter_4" ],
-       "result" : ["Multiply_5"],
-       "result_shape" : [ 3, 4 ],
-       "result_type" :
-           {"bitwidth" : 32, "c_type_string" : "float", "is_real" : true, "is_signed" : true, "is_quantized" : false}
-    }])";
-    stringstream ss(js);
-
-    shared_ptr<Function> f = ngraph::deserialize(ss);
-
-    // max(x,broadcast(Constant(0)))
-    auto backend = runtime::Backend::create("${BACKEND_NAME}");
-
-    // If this compiles it works
-}
-
 NGRAPH_TEST(${BACKEND_NAME}, function_call)
 {
     // First create "f(A,B,C) = (A+B)*C".
@@ -1371,410 +1294,6 @@ NGRAPH_TEST(${BACKEND_NAME}, function_call)
 
     backend->call_with_validate(g, {result}, {x, z, y});
     EXPECT_EQ((vector<float>{194, 296, 418, 560}), read_vector<float>(result));
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_scalar_vector)
-{
-    Shape shape_a{};
-    auto A = make_shared<op::Parameter>(element::f32, shape_a);
-    Shape shape_r{4};
-    auto f = make_shared<Function>(make_shared<op::Broadcast>(A, shape_r, AxisSet{0}),
-                                   op::ParameterVector{A});
-
-    auto backend = runtime::Backend::create("${BACKEND_NAME}");
-
-    // Create some tensors for input/output
-    auto a = backend->create_tensor(element::f32, shape_a);
-    copy_data(a, vector<float>{6});
-    auto result = backend->create_tensor(element::f32, shape_r);
-
-    backend->call_with_validate(f, {result}, {a});
-    EXPECT_EQ((vector<float>{6, 6, 6, 6}), read_vector<float>(result));
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_to_non_existent_axis)
-{
-    Shape shape_a{};
-    auto A = make_shared<op::Parameter>(element::f32, shape_a);
-    Shape shape_r{4};
-    ASSERT_THROW(auto f = make_shared<Function>(
-                     make_shared<op::Broadcast>(A, shape_r, AxisSet{0, 1}), op::ParameterVector{A}),
-                 ngraph_error);
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_scalar_matrix)
-{
-    Shape shape_a{};
-    auto A = make_shared<op::Parameter>(element::f32, shape_a);
-    Shape shape_r{2, 2};
-    auto f = make_shared<Function>(make_shared<op::Broadcast>(A, shape_r, AxisSet{0, 1}),
-                                   op::ParameterVector{A});
-
-    auto backend = runtime::Backend::create("${BACKEND_NAME}");
-
-    // Create some tensors for input/output
-    auto a = backend->create_tensor(element::f32, shape_a);
-    copy_data(a, vector<float>{6});
-    auto result = backend->create_tensor(element::f32, shape_r);
-
-    backend->call_with_validate(f, {result}, {a});
-    EXPECT_EQ((vector<float>{6, 6, 6, 6}), read_vector<float>(result));
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_scalar_tensor)
-{
-    Shape shape_a{};
-    auto A = make_shared<op::Parameter>(element::f32, shape_a);
-    Shape shape_r{2, 2, 2};
-    auto f = make_shared<Function>(make_shared<op::Broadcast>(A, shape_r, AxisSet{0, 1, 2}),
-                                   op::ParameterVector{A});
-
-    auto backend = runtime::Backend::create("${BACKEND_NAME}");
-
-    // Create some tensors for input/output
-    auto a = backend->create_tensor(element::f32, shape_a);
-    copy_data(a, vector<float>{6});
-    auto result = backend->create_tensor(element::f32, shape_r);
-
-    backend->call_with_validate(f, {result}, {a});
-    EXPECT_EQ((vector<float>{6, 6, 6, 6, 6, 6, 6, 6}), read_vector<float>(result));
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_trivial)
-{
-    Shape shape{2, 2, 2};
-    auto A = make_shared<op::Parameter>(element::f32, shape);
-    auto f = make_shared<Function>(make_shared<op::Broadcast>(A, shape, AxisSet{}),
-                                   op::ParameterVector{A});
-
-    auto backend = runtime::Backend::create("${BACKEND_NAME}");
-
-    // Create some tensors for input/output
-    auto a = backend->create_tensor(element::f32, shape);
-    copy_data(a, vector<float>{2, 4, 6, 8, 16, 32, 64, 128});
-    auto result = backend->create_tensor(element::f32, shape);
-
-    backend->call_with_validate(f, {result}, {a});
-    EXPECT_EQ((vector<float>{2, 4, 6, 8, 16, 32, 64, 128}), read_vector<float>(result));
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_vector_colwise)
-{
-    Shape shape_a{3};
-    auto A = make_shared<op::Parameter>(element::f32, shape_a);
-    Shape shape_r{3, 4};
-    auto f = make_shared<Function>(make_shared<op::Broadcast>(A, shape_r, AxisSet{1}),
-                                   op::ParameterVector{A});
-
-    auto backend = runtime::Backend::create("${BACKEND_NAME}");
-
-    // Create some tensors for input/output
-    auto a = backend->create_tensor(element::f32, shape_a);
-    copy_data(a, vector<float>{1, 2, 3});
-    auto result = backend->create_tensor(element::f32, shape_r);
-
-    backend->call_with_validate(f, {result}, {a});
-    EXPECT_EQ((vector<float>{1, 1, 1, 1, 2, 2, 2, 2, 3, 3, 3, 3}), read_vector<float>(result));
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_vector_rowwise)
-{
-    Shape shape_a{4};
-    auto A = make_shared<op::Parameter>(element::f32, shape_a);
-    Shape shape_r{3, 4};
-    auto f = make_shared<Function>(make_shared<op::Broadcast>(A, shape_r, AxisSet{0}),
-                                   op::ParameterVector{A});
-
-    auto backend = runtime::Backend::create("${BACKEND_NAME}");
-
-    // Create some tensors for input/output
-    auto a = backend->create_tensor(element::f32, shape_a);
-    copy_data(a, vector<float>{1, 2, 3, 4});
-    auto result = backend->create_tensor(element::f32, shape_r);
-
-    backend->call_with_validate(f, {result}, {a});
-    EXPECT_EQ((vector<float>{1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4}), read_vector<float>(result));
-}
-
-// Test hybrid mechanism after broadcast
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_vector_rowwise_reversed)
-{
-    Shape shape_a{4};
-    auto A = make_shared<op::Parameter>(element::f32, shape_a);
-    Shape shape_r{3, 4};
-    auto broadcast = make_shared<op::Broadcast>(A, shape_r, AxisSet{0});
-    auto reverse = make_shared<op::Reverse>(broadcast, AxisSet{1});
-    auto f = make_shared<Function>(reverse, op::ParameterVector{A});
-
-    auto backend = runtime::Backend::create("${BACKEND_NAME}");
-
-    // Create some tensors for input/output
-    auto a = backend->create_tensor(element::f32, shape_a);
-    copy_data(a, vector<float>{1, 2, 3, 4});
-    auto result = backend->create_tensor(element::f32, shape_r);
-
-    backend->call_with_validate(f, {result}, {a});
-    EXPECT_EQ((vector<float>{4, 3, 2, 1, 4, 3, 2, 1, 4, 3, 2, 1}), read_vector<float>(result));
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_vector_rowwise_int64)
-{
-    Shape shape_a{4};
-    auto A = make_shared<op::Parameter>(element::i64, shape_a);
-    Shape shape_r{3, 4};
-    auto f = make_shared<Function>(make_shared<op::Broadcast>(A, shape_r, AxisSet{0}),
-                                   op::ParameterVector{A});
-
-    auto backend = runtime::Backend::create("${BACKEND_NAME}");
-
-    // Create some tensors for input/output
-    auto a = backend->create_tensor(element::i64, shape_a);
-    copy_data(a, vector<int64_t>{1, 2, 3, 4});
-    auto result = backend->create_tensor(element::i64, shape_r);
-
-    backend->call_with_validate(f, {result}, {a});
-    EXPECT_EQ((vector<int64_t>{1, 2, 3, 4, 1, 2, 3, 4, 1, 2, 3, 4}), read_vector<int64_t>(result));
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_scalar_to_matrix_int64)
-{
-    Shape shape_a{1};
-    auto A = make_shared<op::Parameter>(element::i64, shape_a);
-    Shape shape_r{3, 1};
-    auto f = make_shared<Function>(make_shared<op::Broadcast>(A, shape_r, AxisSet{0}),
-                                   op::ParameterVector{A});
-
-    auto backend = runtime::Backend::create("${BACKEND_NAME}");
-
-    // Create some tensors for input/output
-    auto a = backend->create_tensor(element::i64, shape_a);
-    copy_data(a, vector<int64_t>{4});
-    auto result = backend->create_tensor(element::i64, shape_r);
-
-    backend->call_with_validate(f, {result}, {a});
-    EXPECT_EQ((vector<int64_t>{4, 4, 4}), read_vector<int64_t>(result));
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_scalar_to_matrix_int32)
-{
-    Shape shape_a{1};
-    auto A = make_shared<op::Parameter>(element::i32, shape_a);
-    Shape shape_r{3, 1};
-    auto f = make_shared<Function>(make_shared<op::Broadcast>(A, shape_r, AxisSet{0}),
-                                   op::ParameterVector{A});
-
-    auto backend = runtime::Backend::create("${BACKEND_NAME}");
-
-    // Create some tensors for input/output
-    auto a = backend->create_tensor(element::i32, shape_a);
-    copy_data(a, vector<int32_t>{4});
-    auto result = backend->create_tensor(element::i32, shape_r);
-
-    backend->call_with_validate(f, {result}, {a});
-    EXPECT_EQ((vector<int32_t>{4, 4, 4}), read_vector<int32_t>(result));
-}
-
-static void broadcast_test_helper(const Shape& shape_a, const Shape& shape_r, const AxisSet& axis)
-{
-    auto A = make_shared<op::Parameter>(element::f32, shape_a);
-
-    vector<float> inp_data(shape_size<const Shape>(shape_a));
-    iota(inp_data.begin(), inp_data.end(), 1);
-
-    auto f =
-        make_shared<Function>(make_shared<op::Broadcast>(A, shape_r, axis), op::ParameterVector{A});
-
-    auto ref_backend = runtime::Backend::create("INTERPRETER");
-    auto wrk_backend = runtime::Backend::create("${BACKEND_NAME}");
-
-    auto wrk_a = wrk_backend->create_tensor(element::f32, shape_a);
-    copy_data(wrk_a, inp_data);
-
-    auto ref_a = ref_backend->create_tensor(element::f32, shape_a);
-    copy_data(ref_a, inp_data);
-
-    auto wrk_result = wrk_backend->create_tensor(element::f32, shape_r);
-    auto ref_result = ref_backend->create_tensor(element::f32, shape_r);
-
-    wrk_backend->call_with_validate(f, {wrk_result}, {wrk_a});
-    ref_backend->call_with_validate(f, {ref_result}, {ref_a});
-    EXPECT_EQ(read_vector<float>(ref_result), read_vector<float>(wrk_result));
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_algo_vector_middle)
-{
-    Shape shape_a{2};
-    Shape shape_r{3, 2, 4};
-    AxisSet axis{0, 2};
-    broadcast_test_helper(shape_a, shape_r, axis);
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_algo_vector_forward_2)
-{
-    Shape shape_a{2};
-    Shape shape_r{3, 2};
-    AxisSet axis{0};
-    broadcast_test_helper(shape_a, shape_r, axis);
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_algo_vector_forward_3)
-{
-    Shape shape_a{2};
-    Shape shape_r{4, 3, 2};
-    AxisSet axis{0, 1};
-    broadcast_test_helper(shape_a, shape_r, axis);
-}
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_algo_vector_forward_4)
-{
-    Shape shape_a{2};
-    Shape shape_r{5, 4, 3, 2};
-    AxisSet axis{0, 1, 2};
-    broadcast_test_helper(shape_a, shape_r, axis);
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_algo_scalar)
-{
-    Shape shape_a{};
-    Shape shape_r{5, 4, 3, 2};
-    AxisSet axis{0, 1, 2, 3};
-    broadcast_test_helper(shape_a, shape_r, axis);
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_algo_vector_backward_2)
-{
-    Shape shape_a{2};
-    Shape shape_r{2, 3};
-    AxisSet axis{1};
-    broadcast_test_helper(shape_a, shape_r, axis);
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_algo_vector_backward_3)
-{
-    Shape shape_a{2};
-    Shape shape_r{2, 3, 4};
-    AxisSet axis{1, 2};
-    broadcast_test_helper(shape_a, shape_r, axis);
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_algo_vector_backward_4)
-{
-    Shape shape_a{2};
-    Shape shape_r{2, 3, 4, 5};
-    AxisSet axis{1, 2, 3};
-    broadcast_test_helper(shape_a, shape_r, axis);
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_algo_matrix_backward_4)
-{
-    Shape shape_a{4, 5};
-    Shape shape_r{2, 3, 4, 5};
-    AxisSet axis{0, 1};
-    broadcast_test_helper(shape_a, shape_r, axis);
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_algo_matrix_stride_1)
-{
-    Shape shape_a{3, 5};
-    Shape shape_r{2, 3, 4, 5};
-    AxisSet axis{0, 2};
-    broadcast_test_helper(shape_a, shape_r, axis);
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_algo_matrix_stride_2)
-{
-    Shape shape_a{3, 4};
-    Shape shape_r{2, 3, 4, 5};
-    AxisSet axis{0, 3};
-    broadcast_test_helper(shape_a, shape_r, axis);
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_algo_matrix_stride_3)
-{
-    Shape shape_a{2, 4};
-    Shape shape_r{2, 3, 4, 5};
-    AxisSet axis{1, 3};
-    broadcast_test_helper(shape_a, shape_r, axis);
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_algo_3d_backward)
-{
-    Shape shape_a{2, 3, 4};
-    Shape shape_r{5, 2, 3, 4};
-    AxisSet axis{0};
-    broadcast_test_helper(shape_a, shape_r, axis);
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_algo_3d_stride_1)
-{
-    Shape shape_a{2, 3, 4};
-    Shape shape_r{2, 5, 3, 4};
-    AxisSet axis{1};
-    broadcast_test_helper(shape_a, shape_r, axis);
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_algo_3d_stride_2)
-{
-    Shape shape_a{2, 3, 4};
-    Shape shape_r{2, 3, 5, 4};
-    AxisSet axis{2};
-    broadcast_test_helper(shape_a, shape_r, axis);
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_matrix_0)
-{
-    Shape shape_a{2, 2};
-    auto A = make_shared<op::Parameter>(element::f32, shape_a);
-    Shape shape_r{2, 2, 2};
-    auto f = make_shared<Function>(make_shared<op::Broadcast>(A, shape_r, AxisSet{0}),
-                                   op::ParameterVector{A});
-
-    auto backend = runtime::Backend::create("${BACKEND_NAME}");
-
-    // Create some tensors for input/output
-    auto a = backend->create_tensor(element::f32, shape_a);
-    copy_data(a, vector<float>{1, 2, 3, 4});
-    auto result = backend->create_tensor(element::f32, shape_r);
-
-    backend->call_with_validate(f, {result}, {a});
-    EXPECT_EQ((vector<float>{1, 2, 3, 4, 1, 2, 3, 4}), read_vector<float>(result));
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_matrix_1)
-{
-    Shape shape_a{2, 2};
-    auto A = make_shared<op::Parameter>(element::f32, shape_a);
-    Shape shape_r{2, 2, 2};
-    auto f = make_shared<Function>(make_shared<op::Broadcast>(A, shape_r, AxisSet{1}),
-                                   op::ParameterVector{A});
-
-    auto backend = runtime::Backend::create("${BACKEND_NAME}");
-
-    // Create some tensors for input/output
-    auto a = backend->create_tensor(element::f32, shape_a);
-    copy_data(a, vector<float>{1, 2, 3, 4});
-    auto result = backend->create_tensor(element::f32, shape_r);
-
-    backend->call_with_validate(f, {result}, {a});
-    EXPECT_EQ((vector<float>{1, 2, 1, 2, 3, 4, 3, 4}), read_vector<float>(result));
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, broadcast_matrix_2)
-{
-    Shape shape_a{2, 2};
-    auto A = make_shared<op::Parameter>(element::f32, shape_a);
-    Shape shape_r{2, 2, 2};
-    auto f = make_shared<Function>(make_shared<op::Broadcast>(A, shape_r, AxisSet{2}),
-                                   op::ParameterVector{A});
-
-    auto backend = runtime::Backend::create("${BACKEND_NAME}");
-
-    // Create some tensors for input/output
-    auto a = backend->create_tensor(element::f32, shape_a);
-    copy_data(a, vector<float>{1, 2, 3, 4});
-    auto result = backend->create_tensor(element::f32, shape_r);
-
-    backend->call_with_validate(f, {result}, {a});
-    EXPECT_EQ((vector<float>{1, 1, 2, 2, 3, 3, 4, 4}), read_vector<float>(result));
 }
 
 NGRAPH_TEST(${BACKEND_NAME}, convert_int32_float32)
@@ -4201,23 +3720,6 @@ NGRAPH_TEST(${BACKEND_NAME}, max_pool_3d)
     {
         EXPECT_TRUE(test::all_close(cpu_results.at(i), int_results.at(i), 1.0e-4f, 1.0e-4f));
     }
-}
-
-NGRAPH_TEST(${BACKEND_NAME}, not)
-{
-    Shape shape{2, 2};
-    auto A = make_shared<op::Parameter>(element::boolean, shape);
-    auto f = make_shared<Function>(make_shared<op::Not>(A), op::ParameterVector{A});
-
-    auto backend = runtime::Backend::create("${BACKEND_NAME}");
-
-    // Create some tensors for input/output
-    auto a = backend->create_tensor(element::boolean, shape);
-    copy_data(a, vector<char>{1, 0, 2, 0});
-    auto result = backend->create_tensor(element::boolean, shape);
-
-    backend->call_with_validate(f, {result}, {a});
-    EXPECT_EQ((vector<char>{0, 1, 0, 1}), read_vector<char>(result));
 }
 
 NGRAPH_TEST(${BACKEND_NAME}, reverse_0d)
